backend/MyProject/requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.0.0
django-cors-headers>= 3.6.0
django-filter>= 2.4.0
psycopg2>= 2.8.6
djangorestframework-simplejwt >=4.6.0
django-cors-headers >= 3.7.0